refactor: get rid of generics
I'm sick of using generics to describe the only two kinds of values that should be used for "T" (f32 and f64). I've refactored this crate to just use a feature flag (`f32`) for this rather than generics. Who in their right mind is mixing f64s and f32s in an analysis anyway? If that is needed, just compile two different programs and be done with it.

This cleans up the code enormously, makes it much more readable, gets rid of the `num` dependency entirely, and improves benchmark performance by up to 80% in some cases.