feat: reorganize, make Bounds a part of functions and individual algorithms
Rather than Bounds being explicitly given when the minimizer is created, they are now part of the function, except in the cases of algorithms which support bounds, in which case they are part of the algorithm. This distinguishes the act of transforming functions to make them bounded and using a algorithm that supports bounding standard functions. A wrapper struct, `BoundedFunction`, is provided to handle the former case.

This also moves tolerances out of terminators and into the main algorithm structs, which is mostly aesthetic but also gets around a few annoyances in coding.