feat: implement Adam algorithm
This is an algorithm intended for stochastic optimization (which is probably one of the reasons it takes so many steps to converge in the non-stochastic tests). It's the de-facto standard for SGD I believe, so it's a useful algorithm to have. Users can easily implement a `CostFunction` which is stochastic if they so desire.