perf: improvements in gradient/Hessian default impls
Added derivatives benchmark. Reducing allocations in the gradient method show 23-61% improvements in both the gradient and Hessian. Additionally, since we already use a central difference gradient, we don't need to also use a central difference Hessian, so we can cut gradient calls in half (which does give us around a 50% additional improvement by itself). However, this reduces accuracy of the Hessian enough to break tests, so I'll probably refactor this and try again.