perf: remove "improved" Hessian for one which is stable and a bit faster
The previous commit improved the Hessian by shifting it and doing half as many gradient calls, but this loses a lot of precision and fails existing tests. I think we would prefer correct Hessians over fast ones, and this rewind includes some small improvements in allocations which gives around 7% improvement in the Hessian in some cases.