Remove `Engine` and add `Input` associated type to `CostFunction` (#77)
* feat: first draft of `Engine`-less implementation

* fix: set f_previous in f-terminator

* bench: remove config and initial position from benchmarks

* style: remove commented-out termination conditions

* refactor: remove `Gradient` and `Hessian` traits

If we are already providing default impls of central finite differences for every `CostFunction`, there is never a situation where we would define a `Gradient`/`Hessian` without defining an underlying function anyway.

* docs: update documentation for `Algorithm` to reflect new generics

* style: allow unused variables in Updatable signature

* feat: add a `Callbacks` structure and some implementations to easily provide default/recommended `Callback`s

* fix: some clippy lints

* test: fix doctests

* feat: remove Updatable trait and unmerge CostFunction and Gradient

this is in anticipation of diversifying callbacks to be able to update problems and also adding parameter types to callbacks

* feat: make `Callback` mutable over `P` to replace `Updatable` trait, consolidate `Callbacks` interface

* feat: add Input associated type to CostFunction

* feat: remodel simulated annealing algorithm to accept a generator-like problem

* test: make test annealing generator more straightforward to use to give future users a better idea of how they should implement their own

* feat: differentiate different kinds of callbacks into terminators, observers, and callbacks

* perf: improvements in gradient/Hessian default impls

Added derivatives benchmark. Reducing allocations in the gradient method show 23-61% improvements in both the gradient and Hessian. Additionally, since we already use a central difference gradient, we don't need to also use a central difference Hessian, so we can cut gradient calls in half (which does give us around a 50% additional improvement by itself). However, this reduces accuracy of the Hessian enough to break tests, so I'll probably refactor this and try again.

* perf: remove "improved" Hessian for one which is stable and a bit faster

The previous commit improved the Hessian by shifting it and doing half as many gradient calls, but this loses a lot of precision and fails existing tests. I think we would prefer correct Hessians over fast ones, and this rewind includes some small improvements in allocations which gives around 7% improvement in the Hessian in some cases.

* fix: revert to previous Hessian implementation which is only about 1.7% slower and will be more accurate for x ~> 0