feat: add `process_default` method to `Algorithm` trait
Introduces the `process_default` method, which processes a problem using `default_callbacks` and configuration. Also updates the `Config` associated type to require `Default` so this method can work in the first place. That last part also makes it easier to use default configs without having to find the config struct.