feat: decouple configs from algorithms and provide `new` methods to avoid invalid state
Previously, if you forgot to set an initial starting point or a set of walkers or a distribution of particles, some algorithms would just panic. We can instead prevent invalid state by refusing to provide a constructor for a config which is invalid, i.e. removing many of the `Default` implementations for configs. Since we always call `with_x0` or similar methods after `default()` anyway, this makes the syntax less verbose and makes it much more clear to the user what is required.