Transform API (#78)
* feat: add `Transform` trait to differentiate explicit bounded algorithms from bounds transforms

This is a potential feature which would simplify and clarify the difference between an algorithm which supports bounds "out of the box" like L-BFGS-B or (as I recently discovered) Nelder-Mead from those which support it via a MINUIIT-style change of coordinates. These changes of coordinates are now a separate trait, and `Bounds` implements this trait. Algorithms which don't natively support bounds can add a bounds transform with a different method from the `with_bounds` method: `with_transform`.

Transformations are also nice because they can transform an entire problem space. For example, when fitting a multivariate normal distribution, symmetric positive-semidefinite covariance matrices are not only desirable, but required. However, with box bounds alone, there is no way to ensure this quality. A coordinate transform (`SymmetricPositiveSemidefiniteTransform`) ensures this property, allowing internal parameters to take any value given by the algorithm in the same way that bounds transforms allow algorithms to work on an unbounded space.

* fix: remove binary reference in Cargo.toml

* test: add tests for transforms/bounds

* test: add tests for wrapped observers/terminators, fix minor inconsistency in MaxSteps, and add message to MaxSteps

* test: remove extraneous code setting parameter names and add test for setting names in minimizations

* style: rename transform methods

* feat: add  and  methods as shorthand for getting owned versions of outputs

* feat: many clippy lints plus an attempt at implementing transforms on L-BFGS-B. However, this will need a refactor, I'm running into some issues with how this currently is set up, but I wanted to commit this just to make sure I have it

* feat: new draft of Transform using higher order derivatives

* feat: make new generic cost/gradient for simulated annealing, remove Input on standard cost function, finish and test transforms, and add demo spherical coordinate transform

* feat: change bounds implementation to one with better inverse, add optional to_internal_jacobian/hessian methods, remove SPD transform, and some perf lints

* fix: accidentally transposed the inverse Jacobian twice

* fix: revert to commented out code from debugging

* feat: move DiffOps into Transform trait and update documentation with new transforms

* feat: add new example generating and fitting multivariate normal data

* feat: add some better styling to multivariate_normal_fit example

* feat: overhaul bounds methods

Big change to make it easier for users to not only implement their own bounds methods, but also to use legacy methods like Minuit/LMFIT bounds. Also a bunch of clippy lints and docstrings.