Development Updates (#82)
* feat: add simultaneous setters for hyperparameters whose bounds depend on each other

This lets us set two parameters like alpha and beta without worrying about an assert panic. For example, if we require alpha < beta and beta defaults to 0.9, then we cannot set alpha to 0.91 even if we plan on later setting beta to 0.95. Rather than check if alpha > beta's default and branching, we can just set them both at the same time.

* fix: unify signatures in Swarm builder methods

These should match the mut self -> Self pattern we use everywhere else

* docs: correct some outdated info in docs

* feat: add Send + Sync to traits that support them

This makes Box<dyn T>'s easier to work with in multithreaded environments

* feat: ensure Some(Float::INFINITY) and Some(Float::NEG_INFINITY) yield unbounded  intervals

Previously, converting (Some(Float::NEG_INFINITY), Some(Float::INFINITY)) to a Bound with Into would create a Bound::LowerAndUpperBound, which is not correct and could cause issues in transforms. This also makes it okay for users to mix and match Nones and infinities.

* feat: add hyperparameter setters for BacktrackingLineSearch

While this line search is currently unused by the provided algorithms, these are useful to have in the event someone wants to build an algorithm which uses it.

* test: add tests for hyperparameter settings in line searches and Nelder-Mead